Add separate Credentials file
