+++ gapic-generator-cloud/templates/cloud/client/_client.erb
@@ -3,7 +3,6 @@
 require "json"
 require "pathname"
 
-require "googleauth"
 require "google/gax"
 <%- if service.client_lro? -%>
 <%= render partial: "client/lro/requires", locals: { service: service } -%>
@@ -11,25 +10,9 @@ require "google/gax"
 
 require "<%= service.gem.version_require %>"
 require "<%= service.service_proto_require %>"
+require "<%= service.credentials_require %>"
 <% end %>
-class <%= service.credentials_class_name %> < Google::Auth::Credentials
-  SCOPE = [
-    "https://www.googleapis.com/auth/cloud-platform"
-  ].freeze
-  PATH_ENV_VARS = %w(<%= service.name.upcase %>_CREDENTIALS
-                     <%= service.name.upcase %>_KEYFILE
-                     GOOGLE_CLOUD_CREDENTIALS
-                     GOOGLE_CLOUD_KEYFILE
-                     GCLOUD_KEYFILE)
-  JSON_ENV_VARS = %w(<%= service.name.upcase %>_CREDENTIALS_JSON
-                     <%= service.name.upcase %>_KEYFILE_JSON
-                     GOOGLE_CLOUD_CREDENTIALS_JSON
-                     GOOGLE_CLOUD_KEYFILE_JSON
-                     GCLOUD_KEYFILE_JSON)
-  DEFAULT_PATHS = ["~/.config/google_cloud/application_default_credentials.json"]
-end
-
-# Service that implements Google Cloud Speech API.
+# Service that implements <%= service.name %> API.
 class <%= service.client_class_name %>
   # @private
   attr_reader :<%= service.stub_name %>
