+++ gapic-generator-cloud/templates/cloud/client/_credentials.erb
@@ -1,0 +1,20 @@
+<%- assert_locals service -%>
+<% @requires = capture do %>
+require "googleauth"
+<% end %>
+class <%= service.credentials_class_name %> < Google::Auth::Credentials
+  SCOPE = [
+    "https://www.googleapis.com/auth/cloud-platform"
+  ].freeze
+  PATH_ENV_VARS = %w(<%= service.name.upcase %>_CREDENTIALS
+                     <%= service.name.upcase %>_KEYFILE
+                     GOOGLE_CLOUD_CREDENTIALS
+                     GOOGLE_CLOUD_KEYFILE
+                     GCLOUD_KEYFILE)
+  JSON_ENV_VARS = %w(<%= service.name.upcase %>_CREDENTIALS_JSON
+                     <%= service.name.upcase %>_KEYFILE_JSON
+                     GOOGLE_CLOUD_CREDENTIALS_JSON
+                     GOOGLE_CLOUD_KEYFILE_JSON
+                     GCLOUD_KEYFILE_JSON)
+  DEFAULT_PATHS = ["~/.config/google_cloud/application_default_credentials.json"]
+end
