+++ gapic-generator-cloud/templates/cloud/credentials.erb
@@ -1,0 +1,6 @@
+<%- assert_locals service -%>
+<%= render partial: "client/credentials",
+           layout: "layouts/ruby",
+           locals: { service: service,
+                     namespaces: service.namespaces }
+%>
